refactor: remove unused tool files and update chat context imports in WorkArea components event drivven messages.
