--- tools/index.ts
@@ -1,5 +1,0 @@
-// tools/index.ts
-import { fetchTool } from "./fetch/fetch.tool";
-import { mathTool } from "./math/math.tool";
-
-export const clientTools = [fetchTool, mathTool];
